:recycle: user record
