+++ src/record/user.rs
@@ -99,3 +99,27 @@ impl NewUserRecord {
         }
     }
 }
+
+#[derive(Debug, Deserialize, Clone, Serialize)]
+pub struct CreateUser {
+    pub id: String,
+    pub loginid: String,
+    pub name: String,
+}
+
+#[derive(Debug, Deserialize, Clone, Serialize)]
+pub struct RequestUser {
+    pub id: String,
+    pub loginid: String,
+    pub name: String,
+}
+
+#[derive(Debug, Deserialize, Clone, Serialize, PartialEq, Eq)]
+pub struct RepresentUser {
+    pub id: String,
+    pub loginid: String,
+    pub name: String,
+}
+
+#[derive(Debug, Deserialize, Clone, Serialize, PartialEq, Eq)]
+pub struct EmptyRepresentUser {}
