+++ src/response.rs
@@ -9,30 +9,6 @@ pub mod title;
 pub mod user;
 
 #[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct CreateUser {
-    pub id: String,
-    pub loginid: String,
-    pub name: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct RequestUser {
-    pub id: String,
-    pub loginid: String,
-    pub name: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize, PartialEq, Eq)]
-pub struct RepresentUser {
-    pub id: String,
-    pub loginid: String,
-    pub name: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize, PartialEq, Eq)]
-pub struct EmptyRepresentUser {}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
 pub struct ProcessCategory {
     pub id: i32,
     pub name: String,
